Migrated bin_ids_forming_sequence from sequences.seqops to new sequences.choices, updating dependencies
