+++ polymerist/rdutils/reactions/reactions.py
@@ -16,7 +16,7 @@ from ..labeling.molwise import ordered_map_nums
 from ..labeling.bondwise import get_bonded_pairs_by_map_nums
 
 from ...genutils.decorators.functional import allow_string_paths, allow_pathlib_paths
-from ...genutils.sequences.seqops import bin_ids_forming_sequence
+from ...genutils.sequences.choices import bin_ids_forming_sequence
 from ...smileslib.substructures import matching_labels_from_substruct_dict
